% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{bar.chart}
\alias{bar.chart}
\title{Sloupcový graf}
\usage{
bar.chart(
  df,
  question,
  poradi_sloupcu = NULL,
  popisky_y = NULL,
  popisky_legenda = NULL,
  moznosti = NULL,
  sloucit = NULL,
  otocit_osy = NULL,
  w = "w",
  nadpis = NULL,
  save = NULL
)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{question}{Nejdelší společná část názvu sloupců, které chceme do grafu vynést (např. "Q2_"), nebo úplný název jediného sloupce, pokud chceme graf s jedním sloupcem (např."Q2_1").}

\item{poradi_sloupcu}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_y, vložíme seznam již s novými názvy.}

\item{popisky_y}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{popisky_legenda}{Volitelný argument, pokud chceme změnit popisky legendy. Stačí vložit dvojce, které chceme změnit v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....)}

\item{moznosti}{Možnosti odpovědí, které chceme zobrazit např. c("Určitě ano","Spíše ano").}

\item{sloucit}{Pokud chceme, aby zvolené možnosti byly zobrazeny dohromady jednou barvou (např. sečíst možnosti určitě ano a spíše ano), vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{otocit_osy}{Pokud chceme, aby sloupce nebyly orientovány horizontálně ale vertikálně vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}

\item{nadpis}{Pokud chceme, aby graf obsahoval znění otázky vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{save}{Pokud chceme, aby se graf automaticky uložil vložíme libovolný text v uvozovkách např. "A", nebo text TRUE. Graf se bude jmenovat: jmeno_otazky.jpeg.}
}
\description{
Funkce vytvoří jednoduchý sloupcový graf, který zobrazuje rozložení vybraných variant odpovědí pro sadu výroků - otázek (otázky ve více sloupcích se společným začátekm názvu). Funkce umožňuje
zvolit si varianty odpovědí, které chceme zobrazit a případně je sloučit (např. určitě ano + spíše ano), měnit popisky u názvů sloupců, otočit osy, nebo seřadit sloupce dle vlastních preferencí. Graf je možné buď automaticky uložit, nebo si upravit velikost dle potřeb a pak jej zkopírovat.
}
\examples{
bar.chart(data1,"d2_", w="VAHA",moznosti = c("Určitě ano","Spíše ano"), nadpis = "A", sloucit = "A",save = "A")
}
