% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\name{bubliny.cross}
\alias{bubliny.cross}
\title{Bubliny cross}
\usage{
bubliny.cross(
  df,
  question,
  rowname_vector,
  poradi_sloupcu_y = NULL,
  poradi_sloupcu_x = NULL,
  popisky_x = NULL,
  popisky_y = NULL,
  moznosti = NULL,
  otocit_osy = NULL,
  w = "w",
  nadpis = NULL,
  save = NULL
)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{question}{Nejdelší společná část názvu sloupců, které chceme do grafu vynést (např. "Q2_"), nebo úplný název jediného sloupce, pokud chceme graf s jednou řadou bublin (např."Q2_1").}

\item{rowname_vector}{Seznam proměnných přes které chceme křížení provést. Tyto budou na ose x. Vložit ve formátu: "jmeno sloupce" nebo c("jméno sloupce 1","jméno sloupce 2",...) - pokud je proměnných více.}

\item{poradi_sloupcu_y}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_y, vložíme seznam již s novými názvy.}

\item{poradi_sloupcu_x}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_x, vložíme seznam již s novými názvy.}

\item{popisky_x}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{popisky_y}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{moznosti}{Možnosti odpovědí, které chceme zobrazit např. c("Určitě ano","Spíše ano").}

\item{otocit_osy}{Pokud chceme prohodit osy, vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}

\item{nadpis}{Pokud chceme, aby graf obsahoval znění otázky vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{save}{Pokud chceme, aby se graf automaticky uložil vložíme libovolný text v uvozovkách např. "A", nebo text TRUE. Graf se bude jmenovat: jmeno_otazky.jpeg.}
}
\description{
Umožňuje zobrazit sadu otázek v několika sloupcích se stejnou sadou odpovědí, kříženou s vybranými dalšími otázkami jako bublinový graf a zobrazit podíl vybraného typu odpovědi nebo jejich kombinaci. Velikost bubliny udává poměr odpovědí pro zvolenou variantu.
}
\examples{
bubliny.cross(data1,"d2_",rowname_vector = c("inte"), w = "VAHA", moznosti = c("Určitě ano","Spíše ano"), otocit_osy = "ano", popisky_x = c("Mám rád Miloše","Mám rád Hroše"))
}
