% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{bubliny.neprave}
\alias{bubliny.neprave}
\title{Bubliny nepravé}
\usage{
bubliny.neprave(
  df,
  question,
  rowname_vektor,
  poradi_sloupcu = NULL,
  popisky_x = NULL,
  popisky_y = NULL,
  w = "w",
  nadpis = NULL,
  save = NULL
)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{question}{Název jednoho slooupce ve formátu: "q12", pro který chceme křížit další proměnné.}

\item{rowname_vektor}{Seznam proměnných přes které chceme křížení provést. Tyto budou na ose x. Vložit ve formátu: "jmeno sloupce" nebo c("jméno sloupce 1","jméno sloupce 2",...) - pokud je proměnných více.}

\item{poradi_sloupcu}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_x, vložíme seznam již s novými názvy.}

\item{popisky_x}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{popisky_y}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}

\item{nadpis}{Pokud chceme, aby graf obsahoval znění otázky vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{save}{Pokud chceme, aby se graf automaticky uložil vložíme libovolný text v uvozovkách např. "A", nebo text TRUE. Graf se bude jmenovat: jmeno_otazky.jpeg.}
}
\description{
Funkce vytvoří bublinový graf, který zobrazuje rozložení jedné proměnné (otázka v jednom sloupci) vůči libovolnému počtu dalších proměnných (např. demografických).
Zobrazuje při tom všechny možnosti odpovědí a jde tak vlastně o jiný typ zobrazení dat ze Svislý sloupcový graf s doplňkem do 100 \%. Funkce umožňuje měnit popisky na ose x,
nebo seřadit sloupce dle vlastních preferencí. Graf je možné buď automaticky uložit, nebo si upravit velikost dle potřeb a pak jej zkopírovat.
}
\examples{
bubliny.neprave(data1,"mat",c("poh","vzd","age"), w = "VAHA", popisky_y = c("AAA","V zásadě chudí"))
}
