% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{bubliny}
\alias{bubliny}
\title{Bubliny}
\usage{
bubliny(
  df,
  question,
  popisky_x = NULL,
  popisky_y = NULL,
  poradi_sloupcu_y = NULL,
  poradi_sloupcu_x = NULL,
  moznosti = NULL,
  otocit_osy = NULL,
  w = "w",
  nadpis = NULL,
  save = NULL
)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{question}{Nejdelší společná část názvu sloupců, které chceme do grafu vynést (např. "Q2_"), nebo úplný název jediného sloupce, pokud chceme graf s jednou řadou bublin (např."Q2_1").}

\item{popisky_x}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{popisky_y}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{poradi_sloupcu_y}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_y, vložíme seznam již s novými názvy.}

\item{poradi_sloupcu_x}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_x, vložíme seznam již s novými názvy.}

\item{moznosti}{Možnosti odpovědí, které chceme zobrazit např. c("Určitě ano","Spíše ano").}

\item{otocit_osy}{Pokud chceme prohodit osy, vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}

\item{nadpis}{Pokud chceme, aby graf obsahoval znění otázky vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{save}{Pokud chceme, aby se graf automaticky uložil vložíme libovolný text v uvozovkách např. "A", nebo text TRUE. Graf se bude jmenovat: jmeno_otazky.jpeg.}
}
\description{
Umožňuje zobrazit sadu otázek v několika sloupcích se stejnou sadou odpovědí jako bublinový graf a zobrazit podíl vybraného typu odpovědi nebo jejich kombinaci. Velikost bubliny udává poměr odpovědí pro zvolenou variantu.
}
\examples{
bubliny(data1,"d6_",w = "VAHA", poradi_sloupcu_x = c("Policie ČR","Hasiči","Vědecké instituce","Hobby uživatelé","Komerční uživatelé")
bubliny(data,"dá_",w = "VAHA",moznosti = c("Určitě ano","Spíše ano"))
}
