% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{f_euro}
\alias{f_euro}
\title{EU Segmentace index}
\usage{
f_euro(df)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely. Sloupce, které jsou k výpočtu třeba nesmí být konvertovány na faktory.}
}
\description{
Funkce, která vrátí kopii původního datasetu doplněnou o sloupce NESPCR, NESPEU a EU_SEGMENTY. Tyto sloupce obsahují indexy pro EU segmentaci.
}
\examples{
data2 <- f_euro(data) #šipkou vytvoříme kopii dat s doplněným indexem
}
