% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{f_KPMG}
\alias{f_KPMG}
\title{KPMG index}
\usage{
f_KPMG(df, tab = NULL, w = "w")
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely. Sloupce, které jsou k výpočtu třeba nesmí být konvertovány na faktory.}

\item{tab}{Volitelný argument, pokud zůstane prázdný, funkce pouze dopočítá indexy. Pokud vložíme jména sloupců (např. demografické proměnné) ve formátu: "jmeno sloupce" nebo: c("jméno sloupce 1","jméno sloupce 2") vygeneruje se zároveň excelová tabulka s průměry indexu I_KPMG v jednotlivých skupinách.}

\item{w}{Volitený argument, smysluplný pouze při tvorbě excelové tabulky. Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}
}
\description{
Funkce, která vrátí kopii původního datasetu doplněnou o sloupce inow ,ifut,iall, I_KPMG (číslo) a indexred (čtyři skupiny) s hodnotami KPMG indexu. Zároveň do konzole vytiskne celkovou hodnotu indexu. Volitelně též vytvoří excelovou tabulku s váženými průměry indexu pro zvolené skupiny a s názvem "KPMG skupiny.xlsx".
}
\examples{
dataX <- f_KPMG(data,tab=c("poh","vzdelani","age"), w = "VAHA")
}
