% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{f_krizeni}
\alias{f_krizeni}
\title{Křížení - excelová tabulka}
\usage{
f_krizeni(df, vektor_radek, vektor_sloupec, w = "w", jmeno_souboru = "krizeni")
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{vektor_radek}{Výčet proměnných, přes které budou ve finální tabulce ve
sloupci ve formátu: "jmeno sloupce" nebo: c("jméno sloupce 1","jméno
sloupce 2") - pokud je proměnných více. Sloupce použité, zde již nemohou
být uvedeny v druhém parametru (před křížením je v takovém případě třeba
vytvořit kopii sloupce s jiným názvem).}

\item{vektor_sloupec}{Výčet proměnných, přes které budeme křížit. Ve finální
tabulce budou v horním řádku. Vložit ve formátu: "jmeno sloupce" nebo:
c("jméno sloupce 1","jméno sloupce 2") - pokud je proměnných více}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný
w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např.
"VÁHA".}

\item{jmeno_souboru}{Pokud chceme, aby se sloupec jmenoval jinak než
"křížení", vložíme název. Např.: "Q1 x Q8".}
}
\description{
Funkce, která vytvoří tebulky s křížením libovolného počtu proměnných na vážených datech a vrátí excelovou tabulku.
}
\examples{
f_krizeni(data,c("mat","inte"),c("poh","vzd","age"), w = "v", jmeno_souboru = "test")


}
