% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{horizontal.bar.complement.multiple}
\alias{horizontal.bar.complement.multiple}
\title{Horizontální sloupcový graf s přehledem možností s doplňkem do 100 \%}
\usage{
horizontal.bar.complement.multiple(
  df,
  question,
  poradi_sloupcu = NULL,
  popisky_y = NULL,
  popisky_legenda = NULL,
  otocit_legendu = FALSE,
  w = "w",
  nadpis = NULL,
  save = NULL
)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{question}{Nejdelší společná část názvu sloupců, které chceme do grafu vynést (např. "Q2_"), nebo úplný název jediného sloupce, pokud chceme graf s jedním sloupcem (např."Q2_1").}

\item{poradi_sloupcu}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_y, vložíme seznam již s novými názvy.}

\item{popisky_y}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{popisky_legenda}{Volitelný argument, pokud chceme změnit popisky legendy. Stačí vložit dvojce, které chceme změnit v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....)}

\item{otocit_legendu}{Otočí pořadí legendy. Stačí vložit libovolný text v uvozovkách např. "A" , nebo text TRUE..}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}

\item{nadpis}{Pokud chceme, aby graf obsahoval znění otázky vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{save}{Pokud chceme, aby se graf automaticky uložil vložíme libovolný text v uvozovkách např. "A", nebo text TRUE. Graf se bude jmenovat: jmeno_otazky.jpeg.}
}
\description{
Funkce vytvoří horizontální sloupcový graf, který zobrazuje rozložení odpovědí pro sadu vybraných výroků, otázek  (otázky ve více sloupcích se společným začátekm názvu). Funkce umožňuje
měnit popisky na ose y, otočit legendu, nebo seřadit sloupce dle vlastních preferencí. Graf je možné buď automaticky uložit, nebo si upravit velikost dle potřeb a pak jej zkopírovat.
}
\examples{
horizontal.bar.complement.multiple(data1,"d2_", w="VAHA",poradi_sloupcu = NULL, otocit_legendu = FALSE, nadpis = TRUE, save = TRUE)
}
