% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{vahy}
\alias{vahy}
\title{Váhy}
\usage{
vahy(df, names, w = 1, tab = NULL)
}
\arguments{
\item{df}{Vstupní dataset.}

\item{names}{Výčet proměnných, na které chceme vážit. Váhy budou použity v pořadí, jaké uvedeme. Vložit ve formátu: "jmeno sloupce" nebo c("jméno sloupce 1","jméno sloupce 2") - pokud je proměnných více. Název sloupce se musí shodovat s názvem sloupce v doprovodné excelové tabulce.}

\item{w}{Volitelný argument. Jsou-li data již vážena a chceme je dále vážit, uvedeme zde název výchozí váhy např.: "v".}

\item{tab}{Chceme-li vytvořit excelovou tabulku s přehledem rozložení libovolných proměnných vložíme názvy těchto proměnných ve formátu: "jmeno sloupce" nebo: c("jméno sloupce 1","jméno sloupce 2",...)}
}
\description{
Vysoce experimentální funkce :-). Vrátí kopii původního datasetu doplněnou o navážená data (název sloupce s váhou - w).
Data budou vážena na proměnné uvedené v argumentu names v daném pořadí. Funkce vyžaduje jako další vstup excelovou tabulku s názvem "vahy.xlsx"
uloženou ve stejné složce jako projekt na kterém pracujete. Tato tabulka nakrmí funkci populačním rozložením pro proměnné na které chceme vážit nebo pro proměnné,
pro které chceme efekt vah sledovat. Tabulka umožňuje rovněž vložení arbitrárních vah (např. pro nevoliče). Funkce rovněž umožňuje vytvořit excelovou tabulku s přehledem efektu vah na data (Vahy přehled.xlsx).
}
\examples{
dataX <- vahy(data3,names =c("el21","vol21"), tab = c("poh","el21","vol21"))
}
