% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkce_vsechny.R
\encoding{UTF-8}
\name{vahy_prehled}
\alias{vahy_prehled}
\title{Přehled rozložení vzorku}
\usage{
vahy_prehled(df, names, w = 1)
}
\arguments{
\item{df}{Vstupní dataset, u kterého chceme zjistit rozložení proměnných.}

\item{names}{Výčet proměnných pro které chceme rozložení zjistit. Vložit ve formátu: "jmeno sloupce" nebo c("jméno sloupce 1","jméno sloupce 2"), je-li proměnných více.}

\item{w}{Volitelný argument. Jsou-li data již vážena a zajímají-li nás již vážené frekvence (např. před dalším vážením).}
}
\description{
Vytvoří excelovou tabulku s přehledem rozložení zvolených proměnných v datech. Tento soubor je možné využít pro úsedek o tom, na které proměnné vážit a po doplnění (a přejmenování na "vahy.xlsx") rovněž jako podklad pro vložení cílových hodnot na které budeme vážit funkcí "vahy". Vytvořený soubor nalezneme pdo názvem "rozlozeni_vzorku.xlsx". Zároveň vrací identický dataframe rozlozeni_vzorku.
}
\examples{
vahy_prehled(data,names =c("el21","vol21","vzd","poh"))
}
