% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_krizeni.R
\encoding{UTF-8}
\name{vertical.bar.complement}
\alias{vertical.bar.complement}
\title{Svislý sloupcový graf s doplňkem do 100 \%}
\usage{
vertical.bar.complement(
  df,
  question,
  rowname_vektor,
  popisky_x = NULL,
  poradi_sloupcu = NULL,
  popisky_legenda = NULL,
  otocit_legendu = FALSE,
  w = "w",
  nadpis = NULL,
  save = NULL
)
}
\arguments{
\item{df}{Vstupní dataset. Sloupce musí obsahovat labely.}

\item{question}{Název jednoho slooupce ve formátu: "q12", pro který chceme křížit další proměnné.}

\item{rowname_vektor}{Seznam proměnných přes které chceme křížení provést. Tyto budou na ose x. Vložit ve formátu: "jmeno sloupce" nebo c("jméno sloupce 1","jméno sloupce 2",...) - pokud je proměnných více.}

\item{popisky_x}{Volitelný argument, pokud chceme změnit názvy sloupců. Pokud ano, je třeba vložit seznam změn v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....).}

\item{poradi_sloupcu}{Volitelný argument, pokud chceme změnit pořadí sloupců. Pokud ano, je třeba vložit seznam všech (!) názvů. Pokud jsme změnili názvy skrze argument popisky_x, vložíme seznam již s novými názvy.}

\item{popisky_legenda}{Volitelný argument, pokud chceme změnit popisky legendy. Stačí vložit dvojce, které chceme změnit v následujícím formátu: c(nové jméno a, staré jméno a, nové jméno b, staré jméno b .....)}

\item{otocit_legendu}{Otočí pořadí legendy. Stačí vložit libovolný text v uvozovkách např. "A" , nebo text TRUE..}

\item{w}{Pokud chceme jako váhu použít jiný sloupec, než sloupec pojmenovaný w. Vložíme název sloupce, který obsahuje váhu, kterou chceme použít. např."VÁHA".}

\item{nadpis}{Pokud chceme, aby graf obsahoval znění otázky vložíme libovolný text v uvozovkách např. "A", nebo text TRUE.}

\item{save}{Pokud chceme, aby se graf automaticky uložil vložíme libovolný text v uvozovkách např. "A", nebo text TRUE. Graf se bude jmenovat: jmeno_otazky.jpeg.}
}
\description{
Funkce vytvoří svislý sloupcový graf, který zobrazuje rozložení jedné proměnné (otázka v jednom sloupci) vůči libovolnému počtu dalších proměnných (např. demografických). Funkce umožňuje
měnit popisky na ose x, otočit legendu, nebo seřadit sloupce dle vlastních preferencí. Graf je možné buď automaticky uložit, nebo si upravit velikost dle potřeb a pak jej zkopírovat.
}
\examples{
vertical.bar.complement(data,"mat",c("poh","vzd","age"), w = "VAHA", save = TRUE )
}
